Added Loading UI when switching page (#1400)
