+++ app/[lng]/chat/chat.tsx
@@ -341,14 +341,16 @@ export default function Chat(props: Props) {
           }
         >
           <div className="flex items-center space-x-2">
-            <ModelSelector
-              placeholder={t("model")}
-              lng={lng}
-              premium={hasGpt4Access()}
-              avModels={avModels}
-              model={model}
-              setModel={setModel}
-            />
+            <span>
+              <ModelSelector
+                placeholder={t("model")}
+                lng={lng}
+                premium={hasGpt4Access()}
+                avModels={avModels}
+                model={model}
+                setModel={setModel}
+              />
+            </span>
             {messages.length === 1 && (
               <Dialog>
                 <DialogTrigger>
