+++ app/[lng]/generations/view/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 import { Suspense } from "react";
 import GenerationViewPage from "./viewpage";
-import { Skeleton } from "@/components/ui/skeleton";
 import { DefaultLanguageParams } from "@/lib/languages";
+import LoadingUI from "@/components/loading";
 
 export default function ViewPage({
   params,
@@ -10,20 +10,8 @@ export default function ViewPage({
   params: DefaultLanguageParams;
 }) {
   return (
-    <Suspense fallback={LoadingUI()}>
+    <Suspense fallback={<LoadingUI />}>
       <GenerationViewPage params={params} />
     </Suspense>
-  );
-}
-function LoadingUI() {
-  return (
-    <div className="flex h-screen w-screen flex-col items-center justify-center">
-      <p>Loading...</p>
-      <div className="space-y-2">
-        <Skeleton className="h-4 w-[250px]" />
-        <Skeleton className="h-4 w-[200px]" />
-        <Skeleton className="h-16 w-[240px]" />
-      </div>
-    </div>
   );
 }
